+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.0] - 2023-07-11
+### Changed
+- Update user authentication to use the new symfony authentication system
+- **Update your custom `config/packages/security.yaml`**
+
 ## [3.1.5] - 2023-01-18
 ### Added
 - Add template filter to pages list view